requirements.txt
asyncio-dgram==2.2.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
Flask==3.1.1
Flask-Limiter==3.12
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcstatus==12.0.1
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.4
rich==13.9.4
setuptools==80.9.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
